chore: give php its own util file
